create payment methods list view
