+++ lib/features/check_out/presentation/widgets/checkout_details.dart
@@ -7,6 +7,7 @@ import 'package:shop_app/core/widgets/custom_buttons.dart';
 
 import '../../../../core/utils/color_manager.dart';
 import 'detials_of_amount.dart';
+import 'payment_method_bottom_sheet_widget.dart';
 
 class CheckoutDetails extends StatelessWidget {
   const CheckoutDetails({super.key, required this.totalPrice});
@@ -61,45 +62,6 @@ class CheckoutDetails extends StatelessWidget {
 }
 
 
-class PaymentMethodsListView extends StatefulWidget {
-  const PaymentMethodsListView({super.key});
-
-  @override
-  State<PaymentMethodsListView> createState() => _PaymentMethodsListViewState();
-}
-
-class _PaymentMethodsListViewState extends State<PaymentMethodsListView> {
-  final List<String> paymentMethodsItems = const [
-    'assets/svg/card.svg',
-    'assets/svg/paypal.svg'
-  ];
-
-  int activeIndex = 0;
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: 62,
-      child: ListView.builder(
-          itemCount: paymentMethodsItems.length,
-          scrollDirection: Axis.horizontal,
-          itemBuilder: (context, index) {
-            return Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 8),
-              child: GestureDetector(
-                onTap: () {
-                  activeIndex = index;
-                  setState(() {});
-                },
-                child: PaymentMethodItem(
-                  isActive: activeIndex == index,
-                  image: paymentMethodsItems[index],
-                ),
-              ),
-            );
-          }),
-    );
-  }
-}
 
 class PaymentMethodItem extends StatelessWidget {
   const PaymentMethodItem({
