+++ lib/features/check_out/presentation/widgets/payment_methods_list_view.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/material.dart';
+
+import 'checkout_details.dart';
+
+class PaymentMethodsListView extends StatefulWidget {
+  const PaymentMethodsListView({super.key});
+
+  @override
+  State<PaymentMethodsListView> createState() => _PaymentMethodsListViewState();
+}
+
+class _PaymentMethodsListViewState extends State<PaymentMethodsListView> {
+  final List<String> paymentMethodsItems = const [
+    'assets/svg/card.svg',
+    'assets/svg/paypal.svg'
+  ];
+
+  int activeIndex = 0;
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 62,
+      child: ListView.builder(
+        itemCount: paymentMethodsItems.length,
+        scrollDirection: Axis.horizontal,
+        itemBuilder: (context, index) {
+          return Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 8),
+            child: GestureDetector(
+              onTap: () {
+                activeIndex = index;
+                setState(() {});
+              },
+              child: PaymentMethodItem(
+                isActive: activeIndex == index,
+                image: paymentMethodsItems[index],
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
